+++ examples/Robot1/BipartiteHungarian.h
@@ -6,10 +6,10 @@
 struct WeightedBipartiteEdge {
     int left;
     int right;
-    int cost;
+    float cost;
 
     WeightedBipartiteEdge() : left(), right(), cost() {}
-    WeightedBipartiteEdge(int left, int right, int cost) : left(left), right(right), cost(cost) {}
+    WeightedBipartiteEdge(int left, int right, float cost) : left(left), right(right), cost(cost) {}
 };
 
 // Given the number of nodes on each side of the bipartite graph and a list of edges, returns a minimum-weight perfect matching.
